A testing utility to generate a cartisian product of test cases.
The fields in the struct representing a test case are
annotated with the "tc" tag, the value of which should be
a JSON array of values.

The generator takes all of the tagged fields and produces
a test case for each (cartisian) combination of the settings.

testCase := struct {
  expectErr bool  `tc:"[true,false]"`
  param     int   `tc:"[1,2,3]"`
  arr       []int `tc:"[[1,2,3],[4,5]"`
}

for gen := GenerateTestCases(&testCase); gen.Next(); {
  // Run your test.  The testCase will have appropriate values
  // for expectErr and param fields set.
}